copy files from another branch

+++ GeneralAlgebraTeamLabProject/Lib/task7/discrete_logarithm.h
@@ -11,6 +11,6 @@
 /*template<class M>
 void print(const std::string_view rem, const M& mmap);*/
 
-LIB_EXPORT std::vector<int> discreteLogarithm(double a, double b, int p);
+LIB_EXPORT std::vector<long long> discreteLogarithm(long long a, long long b, unsigned long long p);
 
 LIB_EXPORT std::vector<unsigned long long> discreteLogarithm2(long base, long result, long module);
